pyproject.toml
[tool.poetry]
name = "web-scrapper"
version = "0.1.0"
description = ""
authors = ["Shorabh Dhandharia"]

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.26.0"
selenium = "^4.1.0"
python-dotenv = "^0.19.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
autopep8 = "^1.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
